requirements.txt
boto3>=1.34.0
PyPDF2>=3.0.0
pdfplumber>=0.10.0
python-docx>=1.1.0
markdown>=3.5.0
beautifulsoup4>=4.12.0
tqdm>=4.66.0
colorama>=0.4.6
pandas>=2.1.0
openpyxl>=3.1.0
tiktoken>=0.5.0
python-magic>=0.4.27
chardet>=5.2.0